All test pass -pre Implementation of new structure
