+++ src/main/resources/application-localstack.yml
@@ -10,6 +10,9 @@ spring:
   autoconfigure:
     exclude:
       - uk.gov.justice.hmpps.sqs.HmppsSqsConfiguration
+  flyway:
+    locations: classpath:db/migration/dev
+
 hmpps:
   sqs:
     enabled: false
