+++ src/main/resources/application.yml
@@ -2,9 +2,30 @@ info.app:
   name: HMPPS Managing Prisoner Apps Api
   version: 1.0
 
+hmpps:
+  service:
+    name: hmpps-managing-prisoner-apps
+    sar:
+      template:
+        path: /sar/templates/sar_template.mustache
+        enabled: true
+  merge:
+    page-size: 50
+  sqs:
+    enabled: false  # Disabled - cloud dev(values.yml) overrides with HMPPS_SQS_ENABLED=true
+    region: eu-west-2
+    queues:
+      domaineventsqueue:
+        queueName: ${HMPPS_SQS_QUEUES_DOMAINEVENTSQUEUE_QUEUE_NAME:}
+        dlqName: ${HMPPS_SQS_QUEUES_DOMAINEVENTSQUEUE_DLQ_NAME:}
+
 spring:
   application:
     name: hmpps-managing-prisoner-apps-api
+  profiles:
+    group:
+      test-sqs:
+        - "localstack"
   codec:
     max-in-memory-size: 10MB
 
@@ -16,6 +37,7 @@ spring:
       hibernate:
         c3p0:
           timeout: 20000
+        dialect: org.hibernate.dialect.PostgreSQLDialect
   datasource:
     username: ${DB_USERNAME}
     password: ${DB_PASSWORD}
@@ -30,8 +52,8 @@ spring:
 
   jackson:
     date-format: "yyyy-MM-dd HH:mm:ss"
-    serialization:
-      WRITE_DATES_AS_TIMESTAMPS: false
+      #serialization:
+    #WRITE_DATES_AS_TIMESTAMPS: false
 
   security:
     oauth2:
@@ -61,6 +83,9 @@ spring:
             client-secret: ${example-api.client.secret}
             authorization-grant-type: client_credentials
             scope: read
+  web:
+    error:
+      include-message: always
 
 server:
   port: 8080
@@ -73,8 +98,6 @@ server:
       protocol-header: x-forwarded-proto
       internal-proxies: 10\.\d{1,3}\.\d{1,3}\.\d{1,3}|192\.168\.\d{1,3}\.\d{1,3}|169\.254\.\d{1,3}\.\d{1,3}|127\.\d{1,3}\.\d{1,3}\.\d{1,3}|172\.1[6-9]{1}\.\d{1,3}\.\d{1,3}|172\.2[0-9]{1}\.\d{1,3}\.\d{1,3}|172\.3[0-1]{1}\.\d{1,3}\.\d{1,3}|0:0:0:0:0:0:0:1|::1|100\.6[4-9]\.\d{1,3}\.\d{1,3}|100\.[7-9][0-9]{1}\.\d{1,3}\.\d{1,3}|100\.1[0-1][0-9]{1}\.\d{1,3}\.\d{1,3}|100\.12[0-7]\.\d{1,3}\.\d{1,3}
   shutdown: graceful
-  error:
-    include-message: always
 
 management:
   endpoints:
