+++ src/test/resources/sar/expected/jpa-entity-schema.json
@@ -1,1 +1,1 @@
-{"App":{"appFiles":"List","appType":"AppType","applicationGroup":"Long","applicationType":"Long","assignedGroup":"UUID","comments":"List","createdBy":"String","createdDate":"LocalDateTime","establishmentId":"String","firstNightCenter":"boolean","genericForm":"boolean","id":"UUID","lastModifiedBy":"String","lastModifiedDate":"LocalDateTime","reference":"String","requestedBy":"String","requestedByFirstName":"String","requestedByLastName":"String","requestedDate":"LocalDateTime","requests":"List","responses":"List","status":"AppStatus"},"AppFile":{"app":"App","createdBy":"String","createdDate":"LocalDateTime","documentId":"String","fileName":"String","fileType":"String","id":"UUID"},"ApplicationGroup":{"applicationTypes":"List","id":"long","name":"String"},"ApplicationType":{"applicationGroup":"ApplicationGroup","genericForm":"boolean","genericType":"boolean","id":"long","logDetailRequired":"boolean","name":"String"},"Comment":{"appId":"UUID","createdBy":"String","createdDate":"LocalDateTime","id":"UUID","message":"String"},"Establishment":{"appTypes":"Set","blackListedAppGroups":"Set","blackListedAppTypes":"Set","defaultDepartments":"boolean","id":"String","name":"String"},"Groups":{"establishmentId":"String","id":"UUID","initialsApplicationTypes":"List","initialsApps":"List","name":"String","type":"GroupType"},"History":{"activity":"Activity","appId":"UUID","createdBy":"String","createdDate":"LocalDateTime","entityId":"UUID","entityType":"EntityType","establishment":"String","id":"UUID"},"Response":{"createdBy":"String","createdDate":"LocalDateTime","decision":"Decision","id":"UUID","reason":"String"}}
+{"App":{"appFiles":"List","appType":"AppType","applicationGroup":"Long","applicationType":"Long","assignedGroup":"UUID","comments":"List","createdBy":"String","createdDate":"LocalDateTime","establishmentId":"String","firstNightCenter":"boolean","genericForm":"boolean","id":"UUID","lastModifiedBy":"String","lastModifiedDate":"LocalDateTime","reference":"String","requestedBy":"String","requestedByFirstName":"String","requestedByLastName":"String","requestedDate":"LocalDateTime","requests":"List","responses":"List","status":"AppStatus","submittedByType":"SubmittedByType"},"AppFile":{"app":"App","createdBy":"String","createdDate":"LocalDateTime","documentId":"String","fileName":"String","fileType":"String","id":"UUID"},"ApplicationGroup":{"applicationTypes":"List","id":"long","name":"String"},"ApplicationType":{"applicationGroup":"ApplicationGroup","genericForm":"boolean","genericType":"boolean","id":"long","logDetailRequired":"boolean","name":"String"},"Comment":{"appId":"UUID","createdBy":"String","createdDate":"LocalDateTime","id":"UUID","message":"String"},"Establishment":{"appTypes":"Set","blackListedAppGroups":"Set","blackListedAppTypes":"Set","defaultDepartments":"boolean","id":"String","name":"String"},"Groups":{"establishmentId":"String","id":"UUID","initialsApplicationTypes":"List","initialsApps":"List","name":"String","type":"GroupType"},"History":{"activity":"Activity","appId":"UUID","createdBy":"String","createdDate":"LocalDateTime","entityId":"UUID","entityType":"EntityType","establishment":"String","id":"UUID","reference":"String"},"Response":{"createdBy":"String","createdDate":"LocalDateTime","decision":"Decision","id":"UUID","reason":"String"}}
